+++ lib/auth/middleware.go
@@ -29,7 +29,6 @@ import (
 	"net/http"
 	"os"
 	"regexp"
-	"slices"
 	"sync/atomic"
 	"time"
 
@@ -54,17 +53,15 @@ import (
 	"github.com/gravitational/teleport/lib/limiter"
 	"github.com/gravitational/teleport/lib/multiplexer"
 	"github.com/gravitational/teleport/lib/observability/metrics"
-	"github.com/gravitational/teleport/lib/tlsca"
-	"github.com/gravitational/teleport/lib/utils"
 	logutils "github.com/gravitational/teleport/lib/utils/log"
 )
 
 const (
-	// TeleportImpersonateUserHeader is a header that specifies teleport user identity
+	// teleportImpersonateUserHeader is a header that specifies teleport user identity
 	// that the proxy is impersonating.
-	TeleportImpersonateUserHeader = "Teleport-Impersonate-User"
-	// TeleportImpersonateIPHeader is a header that specifies the real user IP address.
-	TeleportImpersonateIPHeader = "Teleport-Impersonate-IP"
+	teleportImpersonateUserHeader = "Teleport-Impersonate-User"
+	// teleportImpersonateIPHeader is a header that specifies the real user IP address.
+	teleportImpersonateIPHeader = "Teleport-Impersonate-IP"
 )
 
 // AccessCacheWithEvents extends the [authclient.AccessCache] interface with [types.Events].
@@ -177,12 +174,20 @@ func NewTLSServer(ctx context.Context, cfg TLSServerConfig) (*TLSServer, error) 
 		oldestSupportedVersion = nil
 	}
 
+	apiServer, err := NewAPIServer(&cfg.APIConfig)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
 	// authMiddleware authenticates request assuming TLS client authentication
 	// adds authentication information to the context
 	// and passes it to the API server
 	authMiddleware := &Middleware{
-		ClusterName:            localClusterName.GetClusterName(),
-		AcceptedUsage:          cfg.AcceptedUsage,
+		Middleware: authz.Middleware{
+			ClusterName:   localClusterName.GetClusterName(),
+			AcceptedUsage: cfg.AcceptedUsage,
+			Handler:       apiServer,
+		},
 		Limiter:                limiter,
 		GRPCMetrics:            grpcMetrics,
 		OldestSupportedVersion: oldestSupportedVersion,
@@ -191,12 +196,6 @@ func NewTLSServer(ctx context.Context, cfg TLSServerConfig) (*TLSServer, error) 
 		},
 	}
 
-	apiServer, err := NewAPIServer(&cfg.APIConfig)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	authMiddleware.Wrap(apiServer)
 	// Wrap sets the next middleware in chain to the authMiddleware
 	limiter.WrapHandle(authMiddleware)
 
@@ -321,25 +320,14 @@ func (t *TLSServer) Serve() error {
 
 // Middleware is authentication middleware checking every request
 type Middleware struct {
+	authz.Middleware
+	// TODO(tross): Delete ClusterName once e no longer references it.
 	ClusterName string
-	// Handler is HTTP handler called after the middleware checks requests
-	Handler http.Handler
-	// AcceptedUsage restricts authentication
-	// to a subset of certificates based on certificate metadata,
-	// for example middleware can reject certificates with mismatching usage.
-	// If empty, will only accept certificates with non-limited usage,
-	// if set, will accept certificates with non-limited usage,
-	// and usage exactly matching the specified values.
-	AcceptedUsage []string
 	// Limiter is a rate and connection limiter
 	Limiter *limiter.Limiter
 	// GRPCMetrics is the configured gRPC metrics for the interceptors
 	GRPCMetrics *grpcprom.ServerMetrics
-	// EnableCredentialsForwarding allows the middleware to receive impersonation
-	// identity from the client if it presents a valid proxy certificate.
-	// This is used by the proxy to forward the identity of the user who
-	// connected to the proxy to the next hop.
-	EnableCredentialsForwarding bool
+
 	// OldestSupportedVersion optionally allows the middleware to reject any connections
 	// originated from a client that is using an unsupported version. If not set, then no
 	// rejection occurs.
@@ -351,11 +339,6 @@ type Middleware struct {
 	// lastRejectedAlertTime is the timestamp at which the last alert
 	// was created in response to rejecting unsupported clients.
 	lastRejectedAlertTime atomic.Int64
-}
-
-// Wrap sets next handler in chain
-func (a *Middleware) Wrap(h http.Handler) {
-	a.Handler = h
 }
 
 func getCustomRate(endpoint string) *limiter.RateSet {
@@ -530,7 +513,7 @@ func (a *Middleware) withAuthenticatedUser(ctx context.Context) (context.Context
 	// credentials.TLSInfo is provided if the grpc server is configured with
 	// credentials.NewTLS.
 	case credentials.TLSInfo:
-		user, err := a.GetUser(info.State)
+		user, err := a.GetUser(ctx, info.State)
 		if err != nil {
 			return nil, trace.Wrap(err)
 		}
@@ -619,103 +602,6 @@ func (a *authenticatedStream) Context() context.Context {
 	return a.ctx
 }
 
-// GetUser returns authenticated user based on request TLS metadata
-func (a *Middleware) GetUser(connState tls.ConnectionState) (authz.IdentityGetter, error) {
-	peers := connState.PeerCertificates
-	if len(peers) > 1 {
-		// when turning intermediaries on, don't forget to verify
-		// https://github.com/kubernetes/kubernetes/pull/34524/files#diff-2b283dde198c92424df5355f39544aa4R59
-		return nil, trace.AccessDenied("access denied: intermediaries are not supported")
-	}
-
-	// with no client authentication in place, middleware
-	// assumes not-privileged Nop role.
-	// it theoretically possible to use bearer token auth even
-	// for connections without auth, but this is not active use-case
-	// therefore it is not allowed to reduce scope
-	if len(peers) == 0 {
-		return authz.BuiltinRole{
-			Role:        types.RoleNop,
-			Username:    string(types.RoleNop),
-			ClusterName: a.ClusterName,
-			Identity:    tlsca.Identity{},
-		}, nil
-	}
-	clientCert := peers[0]
-
-	identity, err := tlsca.FromSubject(clientCert.Subject, clientCert.NotAfter)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	// Since 5.0, teleport TLS certs include the origin teleport cluster in the
-	// subject (identity). Before 5.0, origin teleport cluster was inferred
-	// from the cert issuer.
-	certClusterName := identity.TeleportCluster
-	if certClusterName == "" {
-		certClusterName, err = tlsca.ClusterName(clientCert.Issuer)
-		if err != nil {
-			logger.WarnContext(context.Background(), "Failed to parse client certificate", "error", err)
-			return nil, trace.AccessDenied("access denied: invalid client certificate")
-		}
-		identity.TeleportCluster = certClusterName
-	}
-	// If there is any restriction on the certificate usage
-	// reject the API server request. This is done so some classes
-	// of certificates issued for kubernetes usage by proxy, can not be used
-	// against auth server. Later on we can extend more
-	// advanced cert usage, but for now this is the safest option.
-	if len(identity.Usage) != 0 && !slices.Equal(a.AcceptedUsage, identity.Usage) {
-		logger.WarnContext(context.Background(), "Restricted certificate rejected while accessing the auth endpoint",
-			"user", identity.Username,
-			"cert_usage", identity.Usage,
-			"acceptable_usage", a.AcceptedUsage,
-		)
-		return nil, trace.AccessDenied("access denied: invalid client certificate")
-	}
-
-	// this block assumes interactive user from remote cluster
-	// based on the remote certificate authority cluster name encoded in
-	// x509 organization name. This is a safe check because:
-	// 1. Trust and verification is established during TLS handshake
-	// by creating a cert pool constructed of trusted certificate authorities
-	// 2. Remote CAs are not allowed to have the same cluster name
-	// as the local certificate authority
-	if certClusterName != a.ClusterName {
-		// make sure that this user does not have system role
-		// the local auth server can not trust remote servers
-		// to issue certificates with system roles (e.g. Admin),
-		// to get unrestricted access to the local cluster
-		systemRole := findPrimarySystemRole(identity.Groups)
-		if systemRole != nil {
-			return authz.RemoteBuiltinRole{
-				Role:        *systemRole,
-				Username:    identity.Username,
-				ClusterName: certClusterName,
-				Identity:    *identity,
-			}, nil
-		}
-		return newRemoteUserFromIdentity(*identity, certClusterName), nil
-	}
-	// code below expects user or service from local cluster, to distinguish between
-	// interactive users and services (e.g. proxies), the code below
-	// checks for presence of system roles issued in certificate identity
-	systemRole := findPrimarySystemRole(identity.Groups)
-	// in case if the system role is present, assume this is a service
-	// agent, e.g. Proxy, connecting to the cluster
-	if systemRole != nil {
-		return authz.BuiltinRole{
-			Role:                  *systemRole,
-			AdditionalSystemRoles: extractAdditionalSystemRoles(identity.SystemRoles),
-			Username:              identity.Username,
-			ClusterName:           a.ClusterName,
-			Identity:              *identity,
-		}, nil
-	}
-	// otherwise assume that is a local role, no need to pass the roles
-	// as it will be fetched from the local database
-	return newLocalUserFromIdentity(*identity), nil
-}
-
 func findPrimarySystemRole(roles []string) *types.SystemRole {
 	for _, role := range roles {
 		systemRole := types.SystemRole(role)
@@ -725,178 +611,6 @@ func findPrimarySystemRole(roles []string) *types.SystemRole {
 		}
 	}
 	return nil
-}
-
-func extractAdditionalSystemRoles(roles []string) types.SystemRoles {
-	var systemRoles types.SystemRoles
-	for _, role := range roles {
-		systemRole := types.SystemRole(role)
-		err := systemRole.Check()
-		if err != nil {
-			// ignore unknown system roles rather than rejecting them, since new unknown system
-			// roles may be present on certs if we rolled back from a newer version.
-			logger.WarnContext(context.Background(), "Ignoring unknown system role", "unknown_role", role)
-			continue
-		}
-		systemRoles = append(systemRoles, systemRole)
-	}
-	return systemRoles
-}
-
-// ServeHTTP serves HTTP requests
-func (a *Middleware) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	if r.TLS == nil {
-		trace.WriteError(w, trace.AccessDenied("missing authentication"))
-		return
-	}
-	user, err := a.GetUser(*r.TLS)
-	if err != nil {
-		trace.WriteError(w, err)
-		return
-	}
-
-	remoteAddr := r.RemoteAddr
-	// If the request is coming from a trusted proxy and the proxy is sending a
-	// TeleportImpersonateHeader, we will impersonate the user in the header
-	// instead of the user in the TLS certificate.
-	// This is used by the proxy to impersonate the end user when making requests
-	// without re-signing the client certificate.
-	impersonateUser := r.Header.Get(TeleportImpersonateUserHeader)
-	if impersonateUser != "" {
-		if !isProxyRole(user) {
-			trace.WriteError(w, trace.AccessDenied("Credentials forwarding is only permitted for Proxy"))
-			return
-		}
-		// If the service is not configured to allow credentials forwarding, reject the request.
-		if !a.EnableCredentialsForwarding {
-			trace.WriteError(w, trace.AccessDenied("Credentials forwarding is not permitted by this service"))
-			return
-		}
-
-		proxyClusterName := user.GetIdentity().TeleportCluster
-		if user, err = a.extractIdentityFromImpersonationHeader(proxyClusterName, impersonateUser); err != nil {
-			trace.WriteError(w, err)
-			return
-		}
-		remoteAddr = r.Header.Get(TeleportImpersonateIPHeader)
-	}
-
-	// If the request is coming from a trusted proxy, we already know the user
-	// and we will impersonate him. At this point, we need to remove the
-	// TeleportImpersonateHeader from the request, otherwise the proxy will
-	// attempt sending the request to upstream servers with the impersonation
-	// header from a fake user.
-	r.Header.Del(TeleportImpersonateUserHeader)
-	r.Header.Del(TeleportImpersonateIPHeader)
-
-	// determine authenticated user based on the request parameters
-	ctx := r.Context()
-	ctx = authz.ContextWithUserCertificate(ctx, certFromConnState(r.TLS))
-	clientSrcAddr, err := utils.ParseAddr(remoteAddr)
-	if err == nil {
-		ctx = authz.ContextWithClientSrcAddr(ctx, clientSrcAddr)
-	}
-	ctx = authz.ContextWithUser(ctx, user)
-	r = r.WithContext(ctx)
-	// set remote address to the one that was passed in the header
-	// this is needed because impersonation reuses the same connection
-	// and the remote address is not updated from 0.0.0.0:0
-	r.RemoteAddr = remoteAddr
-	a.Handler.ServeHTTP(w, r)
-}
-
-// WrapContextWithUser enriches the provided context with the identity information
-// extracted from the provided TLS connection.
-func (a *Middleware) WrapContextWithUser(ctx context.Context, conn utils.TLSConn) (context.Context, error) {
-	// Perform the handshake if it hasn't been already. Before the handshake we
-	// won't have client certs available.
-	if !conn.ConnectionState().HandshakeComplete {
-		if err := conn.HandshakeContext(ctx); err != nil {
-			return nil, trace.ConvertSystemError(err)
-		}
-	}
-
-	return a.WrapContextWithUserFromTLSConnState(ctx, conn.ConnectionState(), conn.RemoteAddr())
-}
-
-// WrapContextWithUserFromTLSConnState enriches the provided context with the identity information
-// extracted from the provided TLS connection state.
-func (a *Middleware) WrapContextWithUserFromTLSConnState(ctx context.Context, tlsState tls.ConnectionState, remoteAddr net.Addr) (context.Context, error) {
-	user, err := a.GetUser(tlsState)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	ctx = authz.ContextWithUserCertificate(ctx, certFromConnState(&tlsState))
-	ctx = authz.ContextWithClientSrcAddr(ctx, remoteAddr)
-	ctx = authz.ContextWithUser(ctx, user)
-	return ctx, nil
-}
-
-// isProxyRole returns true if the certificate role is a proxy role.
-func isProxyRole(identity authz.IdentityGetter) bool {
-	switch id := identity.(type) {
-	case authz.RemoteBuiltinRole:
-		return id.Role == types.RoleProxy
-	case authz.BuiltinRole:
-		return id.Role == types.RoleProxy
-	default:
-		return false
-	}
-}
-
-// extractIdentityFromImpersonationHeader extracts the identity from the impersonation
-// header and returns it. If the impersonation header holds an identity of a
-// system role, an error is returned.
-func (a *Middleware) extractIdentityFromImpersonationHeader(proxyCluster string, impersonate string) (authz.IdentityGetter, error) {
-	// Unmarshal the impersonated user from the header.
-	var impersonatedIdentity tlsca.Identity
-	if err := json.Unmarshal([]byte(impersonate), &impersonatedIdentity); err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	switch {
-	case findPrimarySystemRole(impersonatedIdentity.Groups) != nil:
-		// make sure that this user does not have system role
-		// since system roles are not allowed to be impersonated.
-		return nil, trace.AccessDenied("can not impersonate a system role")
-	case proxyCluster != "" && proxyCluster != a.ClusterName && proxyCluster != impersonatedIdentity.TeleportCluster:
-		// If a remote proxy is impersonating a user from a different cluster, we
-		// must reject the request. This is because the proxy is not allowed to
-		// impersonate a user from a different cluster.
-		return nil, trace.AccessDenied("can not impersonate users via a different cluster proxy")
-	case impersonatedIdentity.TeleportCluster != a.ClusterName:
-		// if the impersonated user is from a different cluster, we need to
-		// use him as remote user.
-		return newRemoteUserFromIdentity(impersonatedIdentity, impersonatedIdentity.TeleportCluster), nil
-	default:
-		// otherwise assume that is a local role, no need to pass the roles
-		// as it will be fetched from the local database
-		return newLocalUserFromIdentity(impersonatedIdentity), nil
-	}
-}
-
-// newRemoteUserFromIdentity creates a new remote user from the identity.
-func newRemoteUserFromIdentity(identity tlsca.Identity, clusterName string) authz.RemoteUser {
-	return authz.RemoteUser{
-		ClusterName:      clusterName,
-		Username:         identity.Username,
-		Principals:       identity.Principals,
-		KubernetesGroups: identity.KubernetesGroups,
-		KubernetesUsers:  identity.KubernetesUsers,
-		DatabaseNames:    identity.DatabaseNames,
-		DatabaseUsers:    identity.DatabaseUsers,
-		RemoteRoles:      identity.Groups,
-		Identity:         identity,
-	}
-}
-
-// newLocalUserFromIdentity creates a new local user from the identity.
-func newLocalUserFromIdentity(identity tlsca.Identity) authz.LocalUser {
-	return authz.LocalUser{
-		Username: identity.Username,
-		Identity: identity,
-	}
 }
 
 // ImpersonatorRoundTripper is a round tripper that impersonates a user with
@@ -925,14 +639,14 @@ func (r *ImpersonatorRoundTripper) RoundTrip(req *http.Request) (*http.Response,
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
-	req.Header.Set(TeleportImpersonateUserHeader, string(b))
+	req.Header.Set(teleportImpersonateUserHeader, string(b))
 
 	clientSrcAddr, err := authz.ClientSrcAddrFromContext(req.Context())
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
 
-	req.Header.Set(TeleportImpersonateIPHeader, clientSrcAddr.String())
+	req.Header.Set(teleportImpersonateIPHeader, clientSrcAddr.String())
 
 	return r.RoundTripper.RoundTrip(req)
 }
@@ -964,13 +678,13 @@ func IdentityForwardingHeaders(ctx context.Context, originalHeaders http.Header)
 	}
 
 	headers := originalHeaders.Clone()
-	headers.Set(TeleportImpersonateUserHeader, string(b))
+	headers.Set(teleportImpersonateUserHeader, string(b))
 
 	clientSrcAddr, err := authz.ClientSrcAddrFromContext(ctx)
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
 
-	headers.Set(TeleportImpersonateIPHeader, clientSrcAddr.String())
+	headers.Set(teleportImpersonateIPHeader, clientSrcAddr.String())
 	return headers, nil
 }
