+++ lib/srv/db/proxyserver.go
@@ -36,7 +36,6 @@ import (
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/api/types"
 	apiutils "github.com/gravitational/teleport/api/utils"
-	"github.com/gravitational/teleport/lib/auth"
 	"github.com/gravitational/teleport/lib/auth/authclient"
 	"github.com/gravitational/teleport/lib/authz"
 	"github.com/gravitational/teleport/lib/defaults"
@@ -61,7 +60,7 @@ type ProxyServer struct {
 	// cfg is the proxy server configuration.
 	cfg ProxyServerConfig
 	// middleware extracts identity information from client certificates.
-	middleware *auth.Middleware
+	middleware *authz.Middleware
 	// closeCtx is closed when the process shuts down.
 	closeCtx context.Context
 	// log is used for logging.
@@ -165,7 +164,7 @@ func NewProxyServer(ctx context.Context, config ProxyServerConfig) (*ProxyServer
 
 	server := &ProxyServer{
 		cfg: config,
-		middleware: &auth.Middleware{
+		middleware: &authz.Middleware{
 			ClusterName:   clustername.GetClusterName(),
 			AcceptedUsage: []string{teleport.UsageDatabaseOnly},
 		},
